+++ R/conformal_infer.R
@@ -113,6 +113,20 @@ print.int_conformal_full <- function(x, ...) {
   invisible(x)
 }
 
+#' @rdname required_pkgs.int_conformal_cv
+#' @keywords internal
+#' @export
+required_pkgs.int_conformal_full <- function(x, infra = TRUE, ...) {
+  model_pkgs <- required_pkgs(x$wflow, infra = infra)
+
+  if (infra) {
+    model_pkgs <- c(model_pkgs, "probably")
+  }
+  
+  model_pkgs <- unique(model_pkgs)
+  model_pkgs
+}
+
 # ------------------------------------------------------------------------------
 
 #' Prediction intervals from conformal methods