+++ tests/testthat/test-conformal-intervals-split.R
@@ -64,4 +64,13 @@ test_that("split conformal intervals", {
     nrow(new_int),
     nrow(sim_new)
   )
+
+  expect_identical(
+    required_pkgs(lm_int),
+    c(required_pkgs(lm_int$wflow), "probably")
+  )
+  expect_identical(
+    required_pkgs(lm_int, infra = FALSE),
+    required_pkgs(lm_int$wflow, infra = FALSE)
+  )
 })